Audit remediation: CRC, IP_ADDR, APA, path_buffer, language race, and formatting fixes

+++ modules/network/smap_udpbd/src/ministack.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 
-#define IP_ADDR(a, b, c, d) (((uint8_t)(a) << 24) | ((uint8_t)(b) << 16) | ((uint8_t)(c) << 8) | (uint8_t)(d))
+#define IP_ADDR(a, b, c, d) (((uint32_t)(a) << 24) | ((uint32_t)(b) << 16) | ((uint32_t)(c) << 8) | (uint32_t)(d))
 
 static inline uint32_t htonl(uint32_t n)
 {
